Convert view-counter.js to TypeScript
- Convert view-counter.js to view-counter.tsx with proper TypeScript types
- Add ViewData interface for API response structure
- Add ViewCounterProps interface for component props
- Convert fetcher function to properly typed async function
- Add generic type to useSWR<ViewData> for better type safety
- Replace new Number() with cleaner Number() constructor
- Add proper return type annotation to component
